feat: adicionar dialog de confirmação ao cancelar e deletar registro

+++ frontend/src/components/ActionsToolbar.vue
@@ -1,8 +1,11 @@
 <script setup>
 import api from '@/service/api';
+import { confirmation } from '@/utils/dialog';
+import { useConfirm } from 'primevue';
 
 const emit = defineEmits(['reset']);
-const { path, resource } = defineProps(['path', 'resource']);
+const { path, resource, isResourceLoaded } = defineProps(['path', 'resource', 'isResourceLoaded']);
+const confirm = useConfirm();
 
 const save = async () => {
   const { data } = await api.save(path, resource);
@@ -11,13 +14,15 @@ const save = async () => {
 };
 
 const cancel = () => {
-  emit('reset');
+  confirmation(confirm, () => emit('reset'));
 };
 
-const destroy = async () => {
-  const { data } = await api.destroy(path, resource.id);
-  emit('reset');
-  return data;
+const destroy = () => {
+  confirmation(confirm, async () => {
+    const { data } = await api.destroy(path, resource.id);
+    emit('reset');
+    return data;
+  });
 };
 </script>
 
@@ -26,7 +31,7 @@ const destroy = async () => {
     <template #start>
       <Button label="Salvar" icon="pi pi-save" severity="success" class="mr-2" @click="save" />
       <Button label="Cancelar" icon="pi pi-save" severity="warn" class="mr-2" @click="cancel" />
-      <Button label="Deletar" icon="pi pi-trash" severity="danger" @click="destroy" />
+      <Button label="Deletar" icon="pi pi-trash" severity="danger" @click="destroy" :disabled="!isResourceLoaded" />
     </template>
   </Toolbar>
 </template>
